Modify HTML and js
Modified index.html to get rid of the toasts and add an offcanvas element where the 'call' details will be.
Modified script.js to add functionality to get a random call and put the call details in the offcanvas element.
Also added some more interactivity so that the user can 'ask' clarifying questions by clicking a button, which will display some more details of the issue.
Added functionality to the generate ticket button so that when a user clicks it, the ticket info will display in the ticket window.